// client/linup-logic/src/app/services/generated/model/apiPlayersPostRequest.ts
/**
 * API V1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApiPlayersPostRequest { 
    name: string;
    number: number;
    team_id: number;
    position_id: number;
    weekly_position_rank?: number | null;
    weekly_flex_rank?: number | null;
}